Return succeeding exit code for sub command
Wercker executes steps in a `set -e` context. This means it will exit immediately if a pipeline returns a non-zero status. The shell does not exit if the command that fails is part of the command list immediately following a while or until keyword, part of the test in an if statement, part of any command executed in a && or || list.